Task extension class
This class has a method that is used to act upon the returned
boolean value of a crud operation in the db repositories.
It takes two parameters, an item with a message value for both
a success (true) or failure (false) and the current results list.
Since some requests do multiple actions some can be successful
while others can fail.