Setup controllers for pages
Each page controller will inherit from a common controller
where the common view model will be generated since each page
and action will have a lot in common. If a page action needs
additional data for the view model it will be added to the view
model in that specific action.

Also started working on the authentication part in the authcontroller
with a login and register method