Rework of webpack and npm setup
Bought the application theme for bootstrap 4 and installed it
into the project. This included a bit of rework in the current
setup for both js and sass.

Also put down a lot of work in to the major page of the application
which is the forum page. This included setting up the structure
for vue js with vuex and axios http requests to my api. The forum
page now includes a paged list of posts and two sidebars, one to
the left and one to the right. The right one includes a component
used to filter the posts. This is based upon both predefined filters
which the user will be able to administrate from its profile page,
such as excluded users and how many posts to view per page. But it will
also include the possibility to view posts only by certain users,
if the posts include a starting eleven or not or show posts between
certain dates.

The logged in user can also do a couple of actions on the post based
on what it is authorized to do. Users that are administrators can
scissor posts, which is a way or removing them due to bad content.
A user can also vote on a post, as long as they haven't already voted
or if its a post created by them self. They can also reply to posts
and edit posts. However they can only edit posts created by them selves.

They are also able to view url tips, and if they are either a youtube
link or a link to a twitter post they will be embedded. They are also
able to view starting elevens connected to the post, however the modal
body with a starting eleven component isn't developed yet.

They can also view which users have voted on a certain post if there
are votes available.