Complete refactoring of api controllers
Moved all db requests to a seperate db client in order to isolate
the db logic to a separate service. Also moved all api request related
logic to the repositories in order to clean up the api controllers.

Not only did this enhance the structure of the project but it also
seems like it had a positive impact on the overall performance of the
application.